Fix parsing of conjugated words
Temporary fix. Still not properly parsed because the jotoba response doesn't give info about conjugated words.
For example 冷たかった returns the infinitive 冷たい, so word is now stored temporarily until the next valid word is found (た in this case) and then updated.
Mecab would parse it as 冷たかっ and た, which is basically the same result. But mecab could perhaps allow joining them together with the info it gives about what type of word た is.